v.1.3.2 review_1

+++ src/main/java/ru/yandex/practicum/filmorate/interfaces/dal/FilmStorage.java
@@ -11,5 +11,5 @@ public interface FilmStorage {
 
     Film updateFilm(Film film);
 
-    Film getFilm(long filmId);
+    Film getFilm(int filmId);
 }
