+++ src/main/java/ru/yandex/practicum/filmorate/interfaces/dal/UserStorage.java
@@ -11,5 +11,5 @@ public interface UserStorage {
 
     User updateUser(User user);
 
-    User getUser(long userId);
+    User getUser(int userId);
 }
